`csv-import` CLI fixes - load CSV from Moodle data directory [iet:8772481]

+++ bin/csv-import.php
@@ -8,31 +8,48 @@
  */
 define( 'CLI_SCRIPT', true );
 
-define( 'CSV_FILENAME', __DIR__ . '/../example.csv' );
+require_once(__DIR__ . '/bootstrap.php');
+
+auth_ouopenid\bootstrap::vendor_autoload();
+auth_ouopenid\bootstrap::dotenv_load();
+auth_ouopenid\bootstrap::moodle_config();
+
+require_once($CFG->libdir . '/clilib.php');
+
+define( 'CSV_DIR', getenv( 'OUOP_DATA_DIR' ));
+define( 'CSV_FILENAME', '/example.csv' );
 define( 'CSV_HEADING', true );
 define( 'CSV_UNSTRICT', true );
 
-require_once __DIR__ . '/../../../config.php';
-require_once $CFG->libdir . '/clilib.php';
-require_once __DIR__ . '/../../../vendor/autoload.php';
-
 use IET_OU\Moodle\Auth\Ouopenid\Db\User as OuUser;
 
-$csv_file = CSV_FILENAME;
 
 cli_heading('OU-OpenID CSV importer');
-cli_writeln("Filename:  $csv_file");
 
 if ($argc > 1 && $argv[ $argc - 1 ] === '--delete') {
-    OuUser::delete();
-    cli_writeln('User table emptied.');
+    $input = cli_input('Do you really want to delete? Type Y to delete or n to exit', 'n', [ 'n', 'Y' ], $case = true);
+    if ($input === 'Y') {
+        OuUser::delete();
+        cli_writeln('User table emptied. Exiting.');
+    } else {
+        cli_writeln('Cancelled.');
+    }
+    exit;
 }
 
-$count = OuUser::insertFromCsv($csv_file, CSV_HEADING, CSV_UNSTRICT, function ($idx, $user_id) {
+if ($argc > 1) {
+    $csvfile = CSV_DIR . '/' . clean_param($argv[ $argc - 1 ], PARAM_FILE);
+} else {
+    $csvfile = CSV_DIR . CSV_FILENAME;
+}
+
+cli_writeln("Filename:  $csvfile");
+
+$count = OuUser::insertFromCsv($csvfile, CSV_HEADING, CSV_UNSTRICT, function ($idx, $userid) {
     cli_write('.');
 });
 
 cli_writeln("\nUsers inserted:  $count");
 
 
-//End.
+// End.
